Add ability to send metadata to Discord
This actually makes the app functional! Metadata is still hard coded,
but is being stored in the sqlite database "properly". There are still
some issues here, such as me not knowing the borrow checker and working
around it (please check the FIXME comments I've left), and the fact that
we have structs with just a single field differing (we probably want to
look into fixing that somehow).